Full rename and folder restructure
