+++ README.md
@@ -4,5 +4,5 @@ Navigate to the src folder after cloning.
 
 ```
 python setup.py develop
-python teammagicsupergoal/app.py
+python pymdicator/app.py
 ```
