+++ pymdicator/static/css/clean-blog.css
