+++ pymdicator/static/css/cover.css
