+++ pymdicator/templates/momentum.html
