+++ pymdicator/timeseries.py
